@@ -2,12 +2,14 @@
     "name": "justbetter/magento2-akeneo-bundle",
     "description": "Magento2 bundle for extending the Akeneo connector with awesome features.",
     "require": {
-        "php": ">=8.0",
+        "php": ">=8.2",
         "akeneo/module-magento2-connector-community": "*"
     },
     "require-dev": {
         "bitexpert/phpstan-magento": "^0.11.0",
-        "phpstan/phpstan": "^1.10"
+        "magento/magento-coding-standard": "^33.0",
+        "phpstan/phpstan": "^1.10",
+        "rector/rector": "^0.17.13"
     },
     "type": "magento2-module",
     "authors": [
@@ -36,8 +38,9 @@
     },
     "config": {
         "allow-plugins": {
-            "php-http/discovery": true,
-            "magento/composer-dependency-version-audit-plugin": true
+            "dealerdirect/phpcodesniffer-composer-installer": true,
+            "magento/composer-dependency-version-audit-plugin": true,
+            "php-http/discovery": true
         }
     }
 }
