@@ -2,6 +2,7 @@
 
 namespace JustBetter\AkeneoBundle\Block\Adminhtml\Akeneo;
 
+use Exception;
 use Magento\Backend\Helper\Data;
 use Magento\Framework\Module\Manager;
 use Magento\Backend\Block\Template\Context;
@@ -12,48 +13,19 @@ use JustBetter\AkeneoBundle\Block\Adminhtml\Akeneo\Grid as GridOption;
 
 class Grid extends Extended
 {
-    /**
-     * @var Manager
-     */
-    protected $moduleManager;
-
-    /**
-     * @var akeneoFactory
-     */
-    protected $_akeneoFactory;
-
-    /**
-     * @var Status
-     */
-    protected $_status;
-
-    /**
-     * @param Context         $context
-     * @param Data            $backendHelper
-     * @param AkeneoFactory   $AkeneoFactory
-     * @param Status          $status
-     * @param Manager         $moduleManager
-     * @param array           $data
-     *
-     */
     public function __construct(
         Context $context,
         Data $backendHelper,
-        AkeneoFactory $AkeneoFactory,
-        Status $status,
-        Manager $moduleManager,
+        protected AkeneoFactory $akeneoFactory,
+        protected Status $status,
+        protected Manager $moduleManager,
         array $data = []
-    ) {
-        $this->_akeneoFactory = $AkeneoFactory;
-        $this->_status = $status;
-        $this->moduleManager = $moduleManager;
+    )
+    {
         parent::__construct($context, $backendHelper, $data);
     }
 
-    /**
-     * @return void
-     */
-    protected function _construct()
+    protected function _construct(): void
     {
         parent::_construct();
         $this->setId('postGrid');
@@ -64,13 +36,10 @@ class Grid extends Extended
         $this->setVarNameFilter('post_filter');
     }
 
-    /**
-     * @return $this
-     */
-    protected function _prepareCollection()
+    protected function _prepareCollection(): static
     {
         /* @phpstan-ignore-next-line */
-        $collection = $this->_akeneoFactory->create()->getCollection();
+        $collection = $this->akeneoFactory->create()->getCollection();
         $this->setCollection($collection);
 
         parent::_prepareCollection();
@@ -79,17 +48,17 @@ class Grid extends Extended
     }
 
     /**
-     * @return $this
      * @SuppressWarnings(PHPMD.ExcessiveMethodLength)
+     * @throws Exception
      */
-    protected function _prepareColumns()
+    protected function _prepareColumns(): static
     {
         $this->addColumn(
             'id',
             [
-                'header'           => __('ID'),
-                'type'             => 'number',
-                'index'            => 'id',
+                'header' => __('ID'),
+                'type' => 'number',
+                'index' => 'id',
                 'header_css_class' => 'col-id',
                 'column_css_class' => 'col-id',
             ]
@@ -98,9 +67,9 @@ class Grid extends Extended
         $this->addColumn(
             'import',
             [
-                'header'  => __('Type'),
-                'index'   => 'import',
-                'type'    => 'options',
+                'header' => __('Type'),
+                'index' => 'import',
+                'type' => 'options',
                 'options' => GridOption::getOptionArray0(),
             ]
         );
@@ -109,7 +78,7 @@ class Grid extends Extended
             'code',
             [
                 'header' => __('Code'),
-                'index'  => 'code',
+                'index' => 'code',
             ]
         );
 
@@ -117,8 +86,8 @@ class Grid extends Extended
             'entity_id',
             [
                 'header' => __('Magento Entity ID'),
-                'index'  => 'entity_id',
-                'type'   => 'int',
+                'index' => 'entity_id',
+                'type' => 'int',
             ]
         );
 
@@ -126,11 +95,11 @@ class Grid extends Extended
             'created_at',
             [
                 'header' => __('Created'),
-                'index'  => 'created_at',
-                'type'   => 'datetime',
+                'index' => 'created_at',
+                'type' => 'datetime',
             ]
         );
-        
+
         $this->addExportType($this->getUrl('akeneomanager/*/exportCsv', ['_current' => true]), __('CSV'));
         $this->addExportType($this->getUrl('akeneomanager/*/exportExcel', ['_current' => true]), __('Excel XML'));
 
@@ -142,11 +111,7 @@ class Grid extends Extended
         return parent::_prepareColumns();
     }
 
-    
-    /**
-     * @return $this
-     */
-    protected function _prepareMassaction()
+    protected function _prepareMassaction(): static
     {
         $this->setMassactionIdField('id');
         $this->getMassactionBlock()->setFormFieldName('akeneo');
@@ -160,7 +125,7 @@ class Grid extends Extended
             ]
         );
 
-        $statuses = $this->_status->getOptionArray();
+        $statuses = $this->status->getOptionArray();
 
         $this->getMassactionBlock()->addItem(
             'status',
@@ -183,29 +148,26 @@ class Grid extends Extended
         return $this;
     }
 
-    /**
-     * @return string
-     */
-    public function getGridUrl()
+    public function getGridUrl(): string
     {
         return $this->getUrl('akeneomanager/*/index', ['_current' => true]);
     }
 
     /**
-     * @param \JustBetter\AkeneoBundle\Model\Akeneo|\Magento\Framework\Object $row
+     * @param $item
      * @return string
      */
-    public function getRowUrl($row)
+    public function getRowUrl($item): string
     {
         return $this->getUrl(
             'akeneomanager/*/edit',
-            ['id' => $row->getId()]
+            ['id' => $item->getId()]
         );
     }
-    
-    public static function getOptionArray0()
+
+    public static function getOptionArray0(): array
     {
-        $data_array = array();
+        $data_array = [];
         $data_array['family'] = 'family';
         $data_array['attribute'] = 'attribute';
         $data_array['category'] = 'category';
@@ -215,12 +177,13 @@ class Grid extends Extended
         return ($data_array);
     }
 
-    public static function getValueArray0()
+    public static function getValueArray0(): array
     {
-        $data_array=array();
-        foreach (GridOption::getOptionArray0() as $k=> $v) {
-            $data_array[]=array('value'=>$k,'label'=>$v);
+        $data_array = [];
+        foreach (GridOption::getOptionArray0() as $k => $v) {
+            $data_array[] = ['value' => $k, 'label' => $v];
         }
-        return($data_array);
+
+        return ($data_array);
     }
 }
