@@ -3,33 +3,31 @@
 namespace JustBetter\AkeneoBundle\Job;
 
 use Akeneo\Connector\Api\Data\ImportInterface;
+use Magento\Framework\Exception\NoSuchEntityException;
+use Magento\Store\Api\Data\StoreInterface;
 use Magento\Store\Model\StoreManagerInterface;
 use Akeneo\Connector\Model\ResourceModel\Log\Collection;
 
 class SlackMessage
 {
-    protected $store;
+    protected StoreInterface $store;
 
-    public function __construct(StoreManagerInterface $store)
-    {
+    /**
+     * @throws NoSuchEntityException
+     */
+    public function __construct(
+        StoreManagerInterface $store
+    ) {
         $this->store = $store->getStore();
     }
 
-    /**
-     * @return string
-     */
-    public function success()
+    public function success(): string
     {
         return ':white_check_mark: All of today\'s imports in *' . $this->store->getName()
             . '* have been successfully completed.';
     }
 
-    /**
-     * @param Collection $errorLogs
-     * @param Collection $processingLogs
-     * @return string
-     */
-    public function warning(Collection $errorLogs = null, Collection $processingLogs = null)
+    public function warning(Collection $errorLogs = null, Collection $processingLogs = null): string
     {
         $message = ':warning: *Warning!* There\'s a problem with today’s imports in *' . $this->store->getName()
             . "*.\n\n";
@@ -39,15 +37,11 @@ class SlackMessage
         $message .= (isset($processingLogs) && $processingLogs->getData())
             ? $this->logList($processingLogs, ImportInterface::IMPORT_PROCESSING)
             : '';
+
         return $message;
     }
 
-    /**
-     * @param Collection $logs
-     * @param int $status
-     * @return string
-     */
-    protected function logList(Collection $logs, int $status)
+    protected function logList(Collection $logs, int $status): string
     {
         $message = '';
         switch ($status) {
@@ -59,25 +53,18 @@ class SlackMessage
                 break;
         }
         foreach ($logs->getData() as $log) {
-            $message .= $this->formatList(date('H:i:s', strtotime($log['created_at'])), $log['name']);
+            $message .= $this->formatList(date('H:i:s', strtotime((string) $log['created_at'])), $log['name']);
         }
+
         return $message . "\n";
     }
 
-    /**
-     * @return string
-     */
-    public function noImports()
+    public function noImports(): string
     {
         return $this->warning() . 'No imports have been made today.';
     }
 
-    /**
-     * @param string $dateTime
-     * @param string $name
-     * @return string
-     */
-    protected function formatList(string $dateTime, string $name)
+    protected function formatList(string $dateTime, string $name): string
     {
         return '> •  _' . $dateTime . '_ *' . $name . "*\n";
     }
