@@ -4,27 +4,13 @@ namespace JustBetter\AkeneoBundle\Model;
 
 class AkeneoFactory
 {
-    /**
-     * @var \Magento\Framework\ObjectManagerInterface
-     */
-    protected $_objectManager;
-
-    /**
-     * @param \Magento\Framework\ObjectManagerInterface $objectManager
-     */
-    public function __construct(\Magento\Framework\ObjectManagerInterface $objectManager)
-    {
-        $this->_objectManager = $objectManager;
+    public function __construct(
+        protected Akeneo $akeneo
+    ) {
     }
 
-    /**
-     * Create new country model
-     *
-     * @param array $arguments
-     * @return \Magento\Directory\Model\Country
-     */
-    public function create(array $arguments = [])
+    public function create(): Akeneo
     {
-        return $this->_objectManager->create('JustBetter\AkeneoBundle\Model\Akeneo', $arguments, false);
+        return $this->akeneo;
     }
 }
