@@ -2,12 +2,14 @@
 
 namespace JustBetter\AkeneoBundle\Plugin\Helper\Import;
 
+use Akeneo\Connector\Helper\Import\Product as AkeneoProduct;
 use Akeneo\Connector\Helper\Store as StoreHelper;
 use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\Exception\LocalizedException;
 
 class Product
 {
-    protected $codes = null;
+    protected ?array $codes = null;
 
     public function __construct(
         protected ScopeConfigInterface $config,
@@ -15,13 +17,21 @@ class Product
     ) {
     }
 
-    public function beforeCreateTmpTableFromApi($subject, $result, $tableSuffix, $family = null)
+    /**
+     * @throws LocalizedException
+     */
+    public function beforeCreateTmpTableFromApi(
+        AkeneoProduct $subject,
+        array $result,
+        string $tableSuffix,
+        ?string $family = null
+    ): array
     {
         if (is_null($this->codes)) {
             $this->codes = explode(',', (string)$this->config->getValue('akeneo_connector/justbetter/important_attributes'));
         }
 
-        if (!count($this->codes)) {
+        if (!(is_countable($this->codes) ? count($this->codes) : 0)) {
             return [$result, $tableSuffix, $family];
         }
 
