@@ -2,59 +2,28 @@
 
 namespace JustBetter\AkeneoBundle\Block\Adminhtml\System\Config\Form\Field;
 
-use Magento\Eav\Model\Config;
+use Exception;
 use Magento\Backend\Block\Template\Context;
-use Magento\Store\Api\Data\WebsiteInterface;
-use Magento\Tax\Model\TaxClass\Source\Product;
+use Magento\Config\Block\System\Config\Form\Field\FieldArray\AbstractFieldArray;
 use Magento\Framework\Data\Form\Element\Select;
 use Magento\Framework\Data\Form\Element\Factory;
-use Magento\Config\Block\System\Config\Form\Field\FieldArray\AbstractFieldArray;
+use Magento\Tax\Model\TaxClass\Source\Product;
 
-/**
- * Class Tax
- */
 class TaxIdMapping extends AbstractFieldArray
 {
-    /**
-     * This variable contains a Factory
-     *
-     * @var Factory $elementFactory
-     */
-    protected $elementFactory;
-
-    /**
-     * This variable contains a Product
-     *
-     * @var Product $productTaxClassSource
-     */
-    protected $productTaxClassSource;
-
-    /**
-     * Tax constructor.
-     *
-     * @param Context $context
-     * @param Factory $elementFactory
-     * @param Product $productTaxClassSource
-     * @param array $data
-     */
     public function __construct(
         Context $context,
-        Factory $elementFactory,
-        Product $productTaxClassSource,
+        protected Factory $elementFactory,
+        protected Product $productTaxClassSource,
         array $data = []
     ) {
         parent::__construct($context, $data);
-
-        $this->elementFactory        = $elementFactory;
-        $this->productTaxClassSource = $productTaxClassSource;
     }
 
     /**
      * Initialise form fields
-     *
-     * @return void
      */
-    protected function _construct()
+    protected function _construct(): void
     {
         $this->addColumn('akeneo', ['label' => __('Akeneo Attribute Option Code')]);
         $this->addColumn('magento', ['label' => __('Magento')]);
@@ -68,17 +37,14 @@ class TaxIdMapping extends AbstractFieldArray
      * Render array cell for prototypeJS template
      *
      * @param string $columnName
-     *
-     * @return string
-     * @throws \Exception
+     * @throws Exception
      */
-    public function renderCellTemplate($columnName)
+    public function renderCellTemplate($columnName): string
     {
-        if (!in_array($columnName, ['magento']) || !isset($this->_columns[$columnName])) {
+        if ($columnName != 'magento' || !isset($this->_columns[$columnName])) {
             return parent::renderCellTemplate($columnName);
         }
 
-        /** @var array $options */
         $options = [];
 
         if (isset($this->_columns[$columnName])) {
@@ -87,6 +53,7 @@ class TaxIdMapping extends AbstractFieldArray
 
         /** @var Select $element */
         $element = $this->elementFactory->create('select');
+
         $element->setForm($this->getForm())
             ->setName($this->_getCellInputElementName($columnName))
             ->setHtmlId($this->_getCellInputElementId('<%- _id %>', $columnName))
