@@ -2,50 +2,26 @@
 
 namespace JustBetter\AkeneoBundle\Block\Adminhtml\System\Config\Form\Field;
 
-use Magento\Backend\Block\Template\Context;
-use Magento\Customer\Model\ResourceModel\Group\Collection;
 use Akeneo\Connector\Helper\Import\Attribute as AttributeHelper;
-use Magento\Framework\Data\Form\Element\Factory as ElementFactory;
+use Exception;
+use Magento\Backend\Block\Template\Context;
 use Magento\Config\Block\System\Config\Form\Field\FieldArray\AbstractFieldArray;
+use Magento\Customer\Model\ResourceModel\Group\Collection;
+use Magento\Framework\Data\Form\Element\Factory as ElementFactory;
 
-/**
- * Type class
- */
 class Type extends AbstractFieldArray
 {
-    protected $elementFactory;
-    protected $attributeHelper;
-    protected $customerGroup;
-
-    /**
-     * construct function
-     *
-     * @param Context $context
-     * @param ElementFactory $elementFactory
-     * @param AttributeHelper $attributeHelper
-     * @param Collection $customerGroup
-     * @param array $data
-     */
     public function __construct(
         Context $context,
-        ElementFactory $elementFactory,
-        AttributeHelper $attributeHelper,
-        Collection $customerGroup,
+        protected ElementFactory $elementFactory,
+        protected AttributeHelper $attributeHelper,
+        protected Collection $customerGroup,
         array $data = []
     ) {
         parent::__construct($context, $data);
-
-        $this->attributeHelper = $attributeHelper;
-        $this->elementFactory  = $elementFactory;
-        $this->customerGroup = $customerGroup;
     }
 
-    /**
-     * construct function
-     *
-     * @return void
-     */
-    protected function _construct()
+    protected function _construct(): void
     {
         $this->addColumn('pim_type', ['label' => __('Akeneo Price Attribute Code (-EUR)')]);
         $this->addColumn('magento_type', ['label' => __('Magento Customer Group')]);
@@ -55,12 +31,10 @@ class Type extends AbstractFieldArray
     }
 
     /**
-     * renderCellTemplate function
-     *
      * @param string $columnName
-     * @return string
+     * @throws Exception
      */
-    public function renderCellTemplate($columnName)
+    public function renderCellTemplate($columnName): string
     {
         if ($columnName != 'magento_type' || !isset($this->_columns[$columnName])) {
             return parent::renderCellTemplate($columnName);
@@ -78,6 +52,6 @@ class Type extends AbstractFieldArray
             $options
         );
 
-        return str_replace("\n", '', $element->getElementHtml());
+        return str_replace('\n', '', (string) $element->getElementHtml());
     }
 }
