@@ -1,18 +1,13 @@
 <?php
 
-namespace JustBetter\AkeneoBundle\Controller\Adminhtml\akeneo;
+namespace JustBetter\AkeneoBundle\Controller\Adminhtml\Akeneo;
 
 use Magento\Backend\App\Action;
+use Magento\Framework\Controller\Result\Redirect;
 
-/**
- * Class MassDelete
- */
 class MassDelete extends Action
 {
-    /**
-     * @return \Magento\Backend\Model\View\Result\Redirect
-     */
-    public function execute()
+    public function execute(): Redirect
     {
         $itemIds = $this->getRequest()->getParam('akeneo');
         if (!is_array($itemIds) || empty($itemIds)) {
@@ -20,7 +15,7 @@ class MassDelete extends Action
         } else {
             try {
                 foreach ($itemIds as $itemId) {
-                    $post = $this->_objectManager->get('JustBetter\AkeneoBundle\Model\Akeneo')->load($itemId);
+                    $post = $this->_objectManager->get(\JustBetter\AkeneoBundle\Model\Akeneo::class)->load($itemId);
                     $post->delete();
                 }
                 $this->messageManager->addSuccess(
@@ -30,6 +25,7 @@ class MassDelete extends Action
                 $this->messageManager->addError($e->getMessage());
             }
         }
+
         return $this->resultRedirectFactory->create()->setPath('akeneomanager/*/index');
     }
 }
