@@ -1,38 +1,29 @@
 <?php
 
-namespace JustBetter\AkeneoBundle\Controller\Adminhtml\akeneo;
+namespace JustBetter\AkeneoBundle\Controller\Adminhtml\Akeneo;
 
 use Magento\Backend\App\Action;
 use Magento\Backend\App\Action\Context;
+use Magento\Backend\Model\View\Result\Forward;
 use Magento\Backend\Model\View\Result\ForwardFactory;
 
 class NewAction extends Action
 {
-    /**
-     * @var \Magento\Backend\Model\View\Result\ForwardFactory
-     */
-    protected $resultForwardFactory;
-
-    /**
-     * @param Context $context
-     * @param ForwardFactory $resultForwardFactory
-     */
     public function __construct(
         Context $context,
-        ForwardFactory $resultForwardFactory
+        protected ForwardFactory $resultForwardFactory
     ) {
         parent::__construct($context);
-        $this->resultForwardFactory = $resultForwardFactory;
     }
 
     /**
      * Forward to edit
      *
-     * @return \Magento\Backend\Model\View\Result\Forward
+     * @return Forward
      */
-    public function execute()
+    public function execute(): Forward
     {
-        /** @var \Magento\Backend\Model\View\Result\Forward $resultForward */
+        /** @var Forward $resultForward */
         $resultForward = $this->resultForwardFactory->create();
         return $resultForward->forward('edit');
     }
@@ -40,7 +31,7 @@ class NewAction extends Action
     /**
      * {@inheritdoc}
      */
-    protected function _isAllowed()
+    protected function _isAllowed(): bool
     {
         return true;
     }
