@@ -8,11 +8,9 @@ use JustBetter\AkeneoBundle\Job\ImportMetricUnits as ImportMetricUnitsJob;
 
 class ImportMetricUnits implements ObserverInterface
 {
-    protected ImportMetricUnitsJob $job;
-
-    public function __construct(ImportMetricUnitsJob $job)
-    {
-        $this->job = $job;
+    public function __construct(
+        protected ImportMetricUnitsJob $job
+    ) {
     }
 
     public function execute(Observer $observer): void
