@@ -2,15 +2,15 @@
 
 namespace JustBetter\AkeneoBundle\Plugin\Helper\Import;
 
-use Magento\Framework\DB\Adapter\AdapterInterface;
-use Magento\Store\Model\ScopeInterface as scope;
-use Akeneo\Connector\Helper\Store as StoreHelper;
-use Magento\Framework\App\Config\ScopeConfigInterface;
 use Akeneo\Connector\Helper\Import\Attribute as AttributeHelper;
+use Akeneo\Connector\Helper\Import\Entities as AkeneoEntities;
+use Akeneo\Connector\Helper\Store as StoreHelper;
+use Magento\Store\Model\ScopeInterface as scope;
+use Magento\Framework\App\Config\ScopeConfigInterface;
+use Magento\Framework\DB\Adapter\AdapterInterface;
 
 class Entities
 {
-
     public function __construct(
         protected ScopeConfigInterface $config,
         protected StoreHelper $storeHelper,
@@ -18,34 +18,33 @@ class Entities
     ) {
     }
 
-    public function afterFormatMediaName($subject, $result)
+    public function afterFormatMediaName(AkeneoEntities $subject, $result)
     {
         $extensionEnabled = $this->config->getValue('akeneo_connector/justbetter/formatmedianame', scope::SCOPE_WEBSITE);
         if (!$extensionEnabled) {
             return $result;
         }
 
-        return str_replace("_", "-", $result);
+        return str_replace('_', '-', (string) $result);
     }
 
     /**
      * Before setting the values we use Tax Type value from Akeneo when available
-     *
-     * @param $subject
-     * @param $jobCode
-     * @param $entityTable
-     * @param $data
-     * @param $entityTypeId
-     * @param $storeId
-     * @param $mode
-     * @return array
      */
-    public function beforesetValues($subject, $jobCode, $entityTable, $data, $entityTypeId, $storeId, $mode = AdapterInterface::INSERT_ON_DUPLICATE)
+    public function beforeSetValues(
+        AkeneoEntities $subject,
+        string $jobCode,
+        string $entityTable,
+        array $data,
+        int $entityTypeId,
+        int $storeId,
+        int $mode = AdapterInterface::INSERT_ON_DUPLICATE
+    ): array
     {
-        $additonalTypes = $this->attributeHelper->getAdditionalTypes();
+        $additionalTypes = $this->attributeHelper->getAdditionalTypes();
 
-        foreach ($additonalTypes as $key => $additonalType) {
-            if ($additonalType ===  'tax' && isset($data[$key])) {
+        foreach ($additionalTypes as $key => $additionalType) {
+            if ($additionalType ===  'tax' && isset($data[$key])) {
                 if (isset($data['tax_class_id']) && $data['tax_class_id'] instanceof \Zend_Db_Expr) {
                     $defaultTaxClassId = $data['tax_class_id']->__toString();
                     $data['tax_class_id'] = new \Zend_Db_Expr(
@@ -58,6 +57,7 @@ class Entities
                 }
             }
         }
+
         return [$jobCode, $entityTable, $data, $entityTypeId, $storeId, $mode];
     }
 }
